requirements.txt
pillow
customtkinter
pyqt5          # si usas Qt; si no, bórralo
python-dotenv
